chore(ci+hooks): add check-ci workflows and keys sanity guard
