+++ .husky/pre-commit
@@ -1,6 +1,18 @@
-#!/bin/sh
-. "$(dirname "$0")/_/husky.sh"
+export PANTHEON_DISABLE=1
+export CI=1
 
-pnpm store:commit-memory >/dev/null 2>&1 || true
-pnpm validate:sigillins:changed
-pnpm lint-staged
+if [ -n "$UM_SKIP_HEAVY_HOOKS" ]; then
+  echo "[husky] UM_SKIP_HEAVY_HOOKS=1 → running lint-staged only"
+  pnpm lint-staged || exit $?
+  exit 0
+fi
+
+pnpm lint-staged || exit $?
+pnpm typecheck || exit $?
+pnpm test:unit || exit $?
+pnpm schema:validate || exit $?
+pnpm maps:validate || exit $?
+pnpm repomap:build || exit $?
+pnpm repomap:validate || exit $?
+pnpm sanity || exit $?
+pnpm policy:check || exit $?
