+++ .husky/pre-push
@@ -1,0 +1,5 @@
+export PANTHEON_DISABLE=1
+export CI=1
+
+pnpm test:unit:coverage || exit $?
+pnpm policy:check || exit $?
