+++ MandalaMap.md
@@ -75,14 +75,13 @@ Repository: unified-mandala
 
 - `.github/` — **active** GitHub workflows & templates
   - Workflow definitions (core, nightly, experimental) and community health files.
-  - Notes: Audit instructions encoded as jobs; isolate non-actionable tasks to stop false CI failures.; Core/extended workflows documented in docs/roadmap/v1.0-stabilization-playbook.md.; Sigillin validator is required on main; mandala-map workflow uploads artifacts without blocking CI.; Repo Sanity builds the UI dist and waits on apps/ui/dist/index.html before running map/audit checks to avoid race conditions.; CI-Core startet einen Docker-basierten NATS-Server und ruft `pnpm nats:doctor` + `pnpm test:jetstream` im type-and-tests-Gate auf (Fraktal57). CI-Core erzeugt zusätzlich Coverage via `pnpm test:unit:coverage` (Artefakt `coverage-vitest`), und `policy-check.yml` veröffentlicht die Sigillin-Governance als separates Artefakt `policy-sigillins`; Python-Jobs installieren `src/adapters/requirements.txt` für NetCDF-Engines und `PANTHEON_DISABLE=1` schaltet QA-Analytics in CI ab.
-- Notes (continued): Governance-Workflow ruft `pnpm schema:validate`, `pnpm maps:validate` und `pnpm policy:check` auf; `pnpm sanity` prüft `ciBehaviour`-Blöcke als Teil der Repo-Sanity-Gates (Fraktal65). Fraktal66 markiert Kyverno/Sigillin als optionale Schritte (`PANTHEON_DISABLE=1`), loggt Skip-Warnungen, wenn Kyverno-CLI oder `sigillins:report` fehlen, und erzeugt Repomap-Fallback-Artefakte, falls `scripts/run-dist.mjs` nicht verfügbar ist.
-  - Links: [CI Core](.github/workflows/ci.core.yml), [Nightly Mirror](.github/workflows/ci.nightly.yml), [Sigillin validate](.github/workflows/sigillin-validate.yml), [Mandala map job](.github/workflows/mandala-map.yml)
+  - Notes: Audit instructions encoded as jobs; isolate non-actionable tasks to stop false CI failures.; Core/extended workflows documented in docs/roadmap/v1.0-stabilization-playbook.md.; Sigillin validator is required on main; mandala-map workflow uploads artifacts without blocking CI.; Repo Sanity builds the UI dist and waits on apps/ui/dist/index.html before running map/audit checks to avoid race conditions.; CI-Core startet einen Docker-basierten NATS-Server und ruft `pnpm nats:doctor` + `pnpm test:jetstream` im type-and-tests-Gate auf (Fraktal57). CI-Core erzeugt zusätzlich Coverage via `pnpm test:unit:coverage` (Artefakt `coverage-vitest`), und `policy-check.yml` veröffentlicht die Sigillin-Governance als separates Artefakt `policy-sigillins`; Python-Jobs installieren `src/adapters/requirements.txt` für NetCDF-Engines und `PANTHEON_DISABLE=1` schaltet QA-Analytics in CI ab.; Governance-Workflow ruft `pnpm schema:validate`, `pnpm maps:validate` und `pnpm policy:check` auf; `pnpm sanity` prüft `ciBehaviour`-Blöcke als Teil der Repo-Sanity-Gates (Fraktal65). Fraktal66 markiert Kyverno/Sigillin als optionale Schritte (`PANTHEON_DISABLE=1`), loggt Skip-Warnungen, wenn Kyverno-CLI oder `sigillins:report` fehlen, und erzeugt Repomap-Fallback-Artefakte, falls `scripts/run-dist.mjs` nicht verfügbar ist.; Neue Workflows `check-ci.yml` und `ci-verify.yml` bündeln die pnpm-Gates (typecheck, unit, coverage, schema/maps, repomap, sanity, policy) mit deaktiviertem Analytics (`PANTHEON_DISABLE=1`); Mini-Sigillin-Validator CLI-Guard stellt `pnpm typecheck` nach Husky-Lauf sicher.
+  - Links: [CI Core](.github/workflows/ci.core.yml), [Nightly Mirror](.github/workflows/ci.nightly.yml), [Sigillin validate](.github/workflows/sigillin-validate.yml), [Mandala map job](.github/workflows/mandala-map.yml), [check:ci workflow](.github/workflows/check-ci.yml), [ci:verify workflow](.github/workflows/ci-verify.yml)
 
 - `.husky/` — **active** Git hooks
-  - Pre-commit automation enforcing lint-staged and formatting policies.
-  - Notes: Keep hooks aligned with pnpm scripts and policy suite.
-  - Links: [Pre-commit](.husky/pre-commit)
+  - Pre-commit and pre-push gates wrapping lint-staged and repository sanity checks.
+  - Notes: Husky pre-commit runs `pnpm check:precommit` (lint-staged, typecheck, unit, schema/maps, repomap build/validate, sanity, policy); pre-push ruft `pnpm check:prepush` (coverage + policy) auf, `UM_SKIP_HEAVY_HOOKS=1` reduziert den Hook auf lint-staged.
+  - Links: [Pre-commit](.husky/pre-commit), [Pre-push](.husky/pre-push)
 
 - `aws/` — **active** AWS deployment stubs
   - Lambda@Edge and related infrastructure definitions aligning with deployment ready stacks.
