+++ analysis/scripts-and-commands.json
@@ -14,7 +14,7 @@
     "test:watch": "vitest",
     "test:unit": "vitest run",
     "test:unit:coverage": "vitest run --coverage",
-    "test:jest": "echo \"(Jest deaktiviert – nutze vitest)\" && exit 0",
+    "test:jest": "echo \"(Jest deaktiviert \u2013 nutze vitest)\" && exit 0",
     "test:jetstream": "vitest run --config vitest.config.ts --reporter=default tests/event-bus/jetstream-bus.test.ts",
     "test:agents": "vitest run",
     "test:adapters": "vitest run src/adapters/tests/decorators.test.ts",
@@ -30,7 +30,6 @@
     "start:services": "pnpm -s build && NODE_ENV=production node scripts/dev-services.mjs --mode=prod",
     "docs:build": "typedoc",
     "docs:auto": "node scripts/generate-api-docs.js",
-    "prepare": "husky install",
     "split:conversations": "node scripts/split-conversations.js",
     "grep:conversations": "node scripts/filter-conversations.js",
     "analyze:conversations": "node scripts/analyze-conversations.js",
@@ -85,7 +84,7 @@
     "maps:validate": "node scripts/maps/validate-maps.mjs",
     "schema:validate": "node scripts/validate-schemas.mjs",
     "policy:check": "node scripts/policy-suite.mjs",
-    "sanity": "node scripts/repo-sanity.mjs",
+    "sanity": "node scripts/repo-sanity.mjs && node scripts/sanity/index.mjs",
     "repomap:build": "node scripts/generate-repomap.mjs",
     "repomap:validate": "node scripts/validate-repomap.mjs",
     "map:mandala": "node scripts/mandala-map-validate.mjs",
@@ -126,15 +125,17 @@
     "test:py:all": "node scripts/run-pytest.mjs -q",
     "test:sigil": "node scripts/run-dist.mjs scripts/validate-sigil-cli.ts",
     "test:ui": "vitest run --passWithNoTests",
-    "ci:verify": "pnpm test:ts && pnpm sigils:index:strict && pnpm adapter:build:era5",
+    "ci:verify": "node scripts/ci-verify.mjs",
     "kyverno:validate": "node tools/kyverno-dry-run.mjs",
     "scan:ingest": "node scripts/ingest-external-scan.mjs",
     "adapters:ci:install": "pip install -r src/adapters/requirements.txt",
-    "check:ci": "npx tsc -p tsconfig.json --noEmit && pnpm nats:doctor && pnpm test:ts:ci && pnpm test:jetstream && npx pyright",
+    "check:ci": "pnpm ci:verify",
     "sigillins:scaffold": "node scripts/scaffold-interai-bridges.mjs",
     "sigillins:authoring": "node scripts/sigillin-authoring.mjs",
     "sigillins:report": "node scripts/sigillins-generate-reports.mjs",
     "sigillins:build": "node scripts/build-sigillin-archive.mjs",
-    "trikaya:dashboard": "node scripts/generate-trikaya-dashboard.mjs"
+    "trikaya:dashboard": "node scripts/generate-trikaya-dashboard.mjs",
+    "check:precommit": "pnpm lint-staged && pnpm typecheck && pnpm test:unit && pnpm schema:validate && pnpm maps:validate && pnpm repomap:build && pnpm repomap:validate && pnpm sanity && pnpm policy:check",
+    "check:prepush": "pnpm test:unit:coverage && pnpm policy:check"
   }
 }
