+++ codexfeedback.md
@@ -1,5 +1,7 @@
 # Codex Feedback
 
+- FR-UM-2025-11-10-Fraktal74-HuskyCIGates: Husky-Pre-Commit/-Pre-Push laufen über `pnpm check:precommit`/`pnpm check:prepush` (Bypass via `UM_SKIP_HEAVY_HOOKS=1`), neue Aliase `check:ci`/`ci:verify` rufen `scripts/ci-verify.mjs` auf und GitHub Workflows `check-ci.yml`/`ci-verify.yml` spiegeln die gebündelten Gates mit `PANTHEON_DISABLE=1`; `scripts/sanity/check-keys-empty.mjs` hält `keys/` leer, MandalaMap.\*, Stabilization-Playbook und Command-Catalog dokumentieren den Lauf; `scripts/sigils/mini-sigillin-validator.ts` erhielt einen CLI-Guard (argv-Check + typisierte stdin-Events), sodass `pnpm typecheck` wieder grün läuft.
+
 - FR-UM-2025-11-09-Fraktal73-NATSSetupGuidance: `scripts/setup-dev-env.ps1` erkennt Docker-basierte `nats`-Container, exportiert Status-/Start-Hinweise im InstallReport und markiert laufende Container als erfüllt; README, Community-/PowerShell-Onboarding und `docs/runbooks/nats-jetstream.md` dokumentieren `docker ps -a`, `docker start nats`, `docker logs --tail=50 nats` sowie den Windows-Portcheck `Test-NetConnection 127.0.0.1 -Port 4222`; Command-Catalog, MandalaMap.\* und das Stabilization-Playbook spiegeln die aktualisierte Guidance.
 
 - FR-UM-2025-11-08-Fraktal72-PytestSigillinValidator: `pnpm test:py` nutzt den neuen Wrapper `scripts/run-pytest.mjs` (Auto-.venv + Default-Marker); README, CONTRIBUTING und der Command-Katalog dokumentieren die dist-first Aufrufe. Der Mini-Sigillin-Validator (`pnpm sigils:validate:mini` + Schema + Vitest) prüft CREP/Trikāya/Nächste Schritte; README/Command-Catalog zeigen `--input`/`--json`. MandalaMap.\*, Stabilization-Playbook & codexfeedback\* aktualisieren Fraktal72 (Wrapper + Validator) und halten den Cypress-Hinweis `pnpm exec cypress install` präsent; neu ergänzt wurde `pnpm observability:check`, das Prometheus `/api/v1/targets` und Grafana `/api/health` (Port 3300) prüft und die Monitoring-Doku synchronisiert.
