+++ docs/roadmap/v1.0-stabilization-playbook.md
@@ -33,6 +33,7 @@ Dieses Playbook übersetzt die Ergebnisse des Fraktal40-Entwicklungsgesprächs i
 - _Status 2025-10-27:_ Core-, Extended- und Nightly-Jobs installieren `src/adapters/requirements.txt`, sodass NetCDF4/h5netcdf/scipy offline verfügbar sind; `PANTHEON_DISABLE=1` hält QA-Analytics in CI deaktiviert und Coverage-Jobs greifen auf `pnpm test:unit:coverage` zurück.
 - _Status 2025-10-29:_ CI-Core nutzt die Skript-Aliase `pnpm typecheck`, `pnpm test:unit` und `pnpm test:unit:coverage`; Governance-Workflow ruft `pnpm schema:validate`, `pnpm maps:validate`, `pnpm policy:check` auf und `pnpm sanity` prüft die neuen `ciBehaviour`-Felder.
 - _Status 2025-11-09 (Fraktal73):_ Windows-Setup (`scripts/setup-dev-env.ps1`) erkennt Docker-basierte `nats`-Container, exportiert Status/Start-Hinweise im InstallReport und die NATS-Runbooks dokumentieren `docker ps -a ...`, `docker start nats`, `docker logs --tail=50 nats` sowie den Windows-Portcheck `Test-NetConnection 127.0.0.1 -Port 4222`.
+- _Status 2025-11-10 (Fraktal74):_ Husky-Pre-Commit/-Pre-Push nutzen jetzt `pnpm check:precommit`/`pnpm check:prepush` (Typecheck, Unit, Schema/Maps, Repomap, Sanity, Policy bzw. Coverage + Policy) mit optionalem `UM_SKIP_HEAVY_HOOKS=1`-Bypass; neue PNPM-Aliase `check:ci`/`ci:verify` und die Workflows `check-ci.yml`/`ci-verify.yml` spiegeln die gebündelten Gates, `scripts/ci-verify.mjs` sowie `scripts/sanity/check-keys-empty.mjs` halten Repo-Sanity konsistent; Mini-Sigillin-Validator CLI-Guard behebt den TypeScript-Fehler und hält `pnpm typecheck` grün.
 
 2. **Extended-/Experimental-Schicht**
    - In `ci.experimental.yml` die `|| true`-Konstrukte durch echte Fehlerpfade ersetzen und Guardrails/Agents-Dry-Run nur noch mit optionalen `continue-on-error`-Strategien versehen, sobald die Checks stabil sind.
