+++ package.json
@@ -11,6 +11,7 @@
     "build:agents": "tsc -p tsconfig.agents.json && node -e \"const fs=require('fs'),p='dist/agents/personhood/AIJuristicAgent.js';if(fs.existsSync(p))fs.renameSync(p,p.replace(/\\.js$/,'.cjs'));\"",
     "build": "tsc -p tsconfig.build.json && pnpm build:agents",
     "typecheck": "tsc -p tsconfig.json --noEmit",
+    "postinstall": "husky",
     "build:ui": "pnpm -F mandala-ui build",
     "lint": "pnpm lint:types && pnpm lint:eslint",
     "lint:types": "tsc -p tsconfig.json --noEmit",
@@ -18,6 +19,8 @@
     "format": "prettier --write README.md CONTRIBUTING.md docs/ONBOARDING.md codexfeedback.* scripts/dev-services.mjs",
     "format:check": "prettier --check README.md CONTRIBUTING.md docs/ONBOARDING.md codexfeedback.* scripts/dev-services.mjs",
     "lint-staged": "lint-staged",
+    "check:precommit": "pnpm lint-staged && pnpm typecheck && pnpm test:unit && pnpm schema:validate && pnpm maps:validate && pnpm repomap:build && pnpm repomap:validate && pnpm sanity && pnpm policy:check",
+    "check:prepush": "pnpm test:unit:coverage && pnpm policy:check",
     "test": "vitest run",
     "test:watch": "vitest",
     "test:unit": "vitest run",
@@ -46,7 +49,6 @@
     "start:services": "pnpm -s build && NODE_ENV=production node scripts/dev-services.mjs --mode=prod",
     "docs:build": "typedoc",
     "docs:auto": "node scripts/generate-api-docs.js",
-    "prepare": "husky install",
     "split:conversations": "node scripts/split-conversations.js",
     "grep:conversations": "node scripts/filter-conversations.js",
     "analyze:conversations": "node scripts/analyze-conversations.js",
@@ -101,7 +103,7 @@
     "maps:validate": "node scripts/maps/validate-maps.mjs",
     "schema:validate": "node scripts/validate-schemas.mjs",
     "policy:check": "node scripts/policy-suite.mjs",
-    "sanity": "node scripts/repo-sanity.mjs",
+    "sanity": "node scripts/repo-sanity.mjs && node scripts/sanity/index.mjs",
     "repomap:build": "node scripts/generate-repomap.mjs",
     "repomap:validate": "node scripts/validate-repomap.mjs",
     "map:mandala": "node scripts/mandala-map-validate.mjs",
@@ -142,11 +144,11 @@
     "test:py:all": "node scripts/run-pytest.mjs -q",
     "test:sigil": "node scripts/run-dist.mjs scripts/validate-sigil-cli.ts",
     "test:ui": "vitest run --passWithNoTests",
-    "ci:verify": "pnpm test:ts && pnpm sigils:index:strict && pnpm adapter:build:era5",
+    "ci:verify": "node scripts/ci-verify.mjs",
     "kyverno:validate": "node tools/kyverno-dry-run.mjs",
     "scan:ingest": "node scripts/ingest-external-scan.mjs",
     "adapters:ci:install": "pip install -r src/adapters/requirements.txt",
-    "check:ci": "npx tsc -p tsconfig.json --noEmit && pnpm nats:doctor && pnpm test:ts:ci && pnpm test:jetstream && npx pyright",
+    "check:ci": "pnpm ci:verify",
     "sigillins:scaffold": "node scripts/scaffold-interai-bridges.mjs",
     "sigillins:authoring": "node scripts/sigillin-authoring.mjs",
     "sigillins:report": "node scripts/sigillins-generate-reports.mjs",
